+++ openbas-front/src/admin/components/simulations/simulation/injects/ExerciseInjects.tsx
@@ -1,10 +1,9 @@
 import React, { FunctionComponent, useState } from 'react';
-import { useParams, Link } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { BarChartOutlined, ReorderOutlined } from '@mui/icons-material';
 import { Grid, Paper, ToggleButton, ToggleButtonGroup, Tooltip, Typography } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import * as R from 'ramda';
-import type { Exercise, Inject, InjectTestStatus } from '../../../../../utils/api-types';
+import type { Exercise } from '../../../../../utils/api-types';
 import { ArticleContext, TeamContext } from '../../../common/Context';
 import { useAppDispatch } from '../../../../../utils/hooks';
 import { useHelper } from '../../../../../store';
@@ -26,13 +25,8 @@ import ExerciseDistributionScoreOverTimeByInjectorContract from '../overview/Exe
 import ExerciseDistributionScoreOverTimeByTeam from '../overview/ExerciseDistributionScoreOverTimeByTeam';
 import ExerciseDistributionScoreOverTimeByTeamInPercentage from '../overview/ExerciseDistributionScoreOverTimeByTeamInPercentage';
 import { useFormatter } from '../../../../../components/i18n';
-import useEntityToggle from '../../../../../utils/hooks/useEntityToggle';
-import ToolBar from '../../../common/ToolBar';
-import { isNotEmptyField } from '../../../../../utils/utils';
-import { fetchExerciseInjectsSimple, bulkTestInjects } from '../../../../../actions/injects/inject-action';
-import injectContextForExercise from '../ExerciseContext';
+import { fetchExerciseInjectsSimple } from '../../../../../actions/injects/inject-action';
 import Injects from '../../../common/injects/Injects';
-import { MESSAGING$ } from '../../../../../utils/Environment';
 
 const useStyles = makeStyles(() => ({
   paperChart: {
@@ -76,145 +70,7 @@ const ExerciseInjects: FunctionComponent<Props> = () => {
   const articleContext = articleContextForExercise(exerciseId);
   const teamContext = teamContextForExercise(exerciseId, []);
 
-  const injectContext = injectContextForExercise(exercise);
-
   const [viewMode, setViewMode] = useState('list');
-  const {
-    selectedElements,
-    deSelectedElements,
-    selectAll,
-    handleClearSelectedElements,
-    handleToggleSelectAll,
-    onToggleEntity,
-    numberOfSelectedElements,
-  } = useEntityToggle<{ inject_id: string }>('inject', injects.length);
-  const onRowShiftClick = (currentIndex: number, currentEntity: { inject_id: string }, event: React.SyntheticEvent | null = null) => {
-    if (event) {
-      event.stopPropagation();
-      event.preventDefault();
-    }
-    if (selectedElements && !R.isEmpty(selectedElements)) {
-      // Find the indexes of the first and last selected entities
-      let firstIndex = R.findIndex(
-        (n: Inject) => n.inject_id === R.head(R.values(selectedElements)).inject_id,
-        injects,
-      );
-      if (currentIndex > firstIndex) {
-        let entities: Inject[] = [];
-        while (firstIndex <= currentIndex) {
-          entities = [...entities, injects[firstIndex]];
-          // eslint-disable-next-line no-plusplus
-          firstIndex++;
-        }
-        const forcedRemove = R.values(selectedElements).filter(
-          (n: Inject) => !entities.map((o) => o.inject_id).includes(n.inject_id),
-        );
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-expect-error
-        return onToggleEntity(entities, event, forcedRemove);
-      }
-      let entities: Inject[] = [];
-      while (firstIndex >= currentIndex) {
-        entities = [...entities, injects[firstIndex]];
-        // eslint-disable-next-line no-plusplus
-        firstIndex--;
-      }
-      const forcedRemove = R.values(selectedElements).filter(
-        (n: Inject) => !entities.map((o) => o.inject_id).includes(n.inject_id),
-      );
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-expect-error
-      return onToggleEntity(entities, event, forcedRemove);
-    }
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-expect-error
-    return onToggleEntity(currentEntity, event);
-  };
-
-  const injectsToProcess = selectAll
-    ? injects.filter((inject: Inject) => !R.keys(deSelectedElements).includes(inject.inject_id))
-    : injects.filter(
-      (inject: Inject) => R.keys(selectedElements).includes(inject.inject_id) && !R.keys(deSelectedElements).includes(inject.inject_id),
-    );
-
-  const massUpdateInjects = async (actions: {
-    field: string,
-    type: string,
-    values: { value: string }[]
-  }[]) => {
-    const updateFields = [
-      'inject_title',
-      'inject_description',
-      'inject_injector_contract',
-      'inject_content',
-      'inject_depends_from_another',
-      'inject_depends_duration',
-      'inject_teams',
-      'inject_assets',
-      'inject_asset_groups',
-      'inject_documents',
-      'inject_all_teams',
-      'inject_country',
-      'inject_city',
-      'inject_tags',
-    ];
-    const injectsToUpdate = injectsToProcess.filter((inject: Inject) => inject.inject_injector_contract?.convertedContent);
-    // eslint-disable-next-line no-plusplus
-    for (let i = 0; i < actions.length; i++) {
-      const action = actions[i];
-      // eslint-disable-next-line no-plusplus
-      for (let j = 0; j < injectsToUpdate.length; j++) {
-        const injectToUpdate = {
-          ...injectsToUpdate[j],
-          inject_injector_contract: injectsToUpdate[j].inject_injector_contract.injector_contract_id,
-        };
-        switch (action.type) {
-          case 'ADD':
-            if (isNotEmptyField(injectToUpdate[`inject_${action.field}`])) {
-              injectToUpdate[`inject_${action.field}`] = R.uniq([...injectToUpdate[`inject_${action.field}`], ...action.values.map((n) => n.value)]);
-            } else {
-              injectToUpdate[`inject_${action.field}`] = R.uniq(action.values.map((n) => n.value));
-            }
-            // eslint-disable-next-line no-await-in-loop
-            await injectContext.onUpdateInject(injectToUpdate.inject_id, R.pick(updateFields, injectToUpdate));
-            break;
-          case 'REPLACE':
-            injectToUpdate[`inject_${action.field}`] = R.uniq(action.values.map((n) => n.value));
-            // eslint-disable-next-line no-await-in-loop
-            await injectContext.onUpdateInject(injectToUpdate.inject_id, R.pick(updateFields, injectToUpdate));
-            break;
-          case 'REMOVE':
-            if (isNotEmptyField(injectToUpdate[`inject_${action.field}`])) {
-              injectToUpdate[`inject_${action.field}`] = injectToUpdate[`inject_${action.field}`].filter((n: string) => !action.values.map((o) => o.value).includes(n));
-            } else {
-              injectToUpdate[`inject_${action.field}`] = [];
-            }
-            // eslint-disable-next-line no-await-in-loop
-            await injectContext.onUpdateInject(injectToUpdate.inject_id, R.pick(updateFields, injectToUpdate));
-            break;
-          default:
-            return;
-        }
-      }
-    }
-  };
-  const bulkDeleteInjects = () => {
-    injectContext.onBulkDeleteInjects(injectsToProcess.map((inject: Inject) => inject.inject_id));
-  };
-
-  const massTestInjects = () => {
-    bulkTestInjects(injectsToProcess.map((inject: Inject) => inject.inject_id)).then((result: { data: InjectTestStatus[] }) => {
-      if (numberOfSelectedElements === 1) {
-        MESSAGING$.notifySuccess(t('Inject test has been sent, you can view test logs details on {itsDedicatedPage}.', {
-          itsDedicatedPage: <Link to={`/admin/exercises/${exercise.exercise_id}/tests/${result.data[0].status_id}`}>{t('its dedicated page')}</Link>,
-        }));
-      } else {
-        MESSAGING$.notifySuccess(t('Inject test has been sent, you can view test logs details on {itsDedicatedPage}.', {
-          itsDedicatedPage: <Link to={`/admin/exercises/${exercise.exercise_id}/tests`}>{t('its dedicated page')}</Link>,
-        }));
-      }
-    });
-  };
 
   return (
     <>
@@ -234,24 +90,6 @@ const ExerciseInjects: FunctionComponent<Props> = () => {
               // @ts-expect-error typing
               teamsUsers={exercise.exercise_teams_users}
               setViewMode={setViewMode}
-              onToggleEntity={onToggleEntity}
-              onToggleShiftEntity={onRowShiftClick}
-              handleToggleSelectAll={handleToggleSelectAll}
-              selectedElements={selectedElements}
-              deSelectedElements={deSelectedElements}
-              selectAll={selectAll}
-            />
-            <ToolBar
-              numberOfSelectedElements={numberOfSelectedElements}
-              selectedElements={selectedElements}
-              deSelectedElements={deSelectedElements}
-              selectAll={selectAll}
-              handleClearSelectedElements={handleClearSelectedElements}
-              context="exercise"
-              id={exercise.exercise_id}
-              handleUpdate={massUpdateInjects}
-              handleBulkDelete={bulkDeleteInjects}
-              handleBulkTest={massTestInjects}
             />
           </TeamContext.Provider>
         </ArticleContext.Provider>
